[#27] 요구사항 수정 완료
